LayoutTests/fast/dom/css-set-property-exception-expected.txt

Test for bug 7296.

This test checks to see whether you get exceptions when setting a property with a "bad value". Setting using JavaScript property syntax and with setProperty() should behave the same.

The results below should show success in cases 1, 3, 5, and 7.

It is OK if the order of properties changes from the expected results - IE 6 and Firefox 2 don't agree on it anyway.

Successfully set display to "block"; cssText is now: "top: 0px; bottom: 1px; display: block; ".
Got exception trying to set display to "foobar"; cssText is now: "top: 0px; display: none; bottom: 1px; ".
Successfully set display to ""; cssText is now: "top: 0px; bottom: 1px; ".
Got exception trying to set display to null; cssText is now: "top: 0px; display: none; bottom: 1px; ".
Successfully set display to "block" with setProperty; cssText is now: "top: 0px; bottom: 1px; display: block; ".
Got exception trying to set display to "foobar" with setProperty; cssText is now: "top: 0px; display: none; bottom: 1px; ".
Successfully set display to "" with setProperty; cssText is now: "top: 0px; bottom: 1px; ".
Got exception trying to set display to null with setProperty; cssText is now: "top: 0px; display: none; bottom: 1px; ".
